refactor flips fetching
